refactor: :art: refactor the struction of code
