+++ logic/Gaming/Game.cs
@@ -270,26 +270,6 @@ namespace Gaming
             else
                 return false;
         }
-        public void AllActiveSkillDisabledTemporarily()
-        {
-            if (!gameMap.Timer.IsGaming)
-                return;
-            gameMap.GameObjLockDict[GameObjType.Character].EnterWriteLock();
-            try
-            {
-                foreach (Character player in gameMap.GameObjDict[GameObjType.Character])
-                {
-                    foreach (var activeSkill in player.Occupation.ListOfIActiveSkill)
-                    {
-                        player.SetTimeUntilActiveSkillAvailable(activeSkill, 0);
-                    }
-                }
-            }
-            finally
-            {
-                gameMap.GameObjLockDict[GameObjType.Character].ExitWriteLock();
-            }
-        }
 
         public void AllPlayerUsePassiveSkill()
         {
