+++ logic/Server/GameServer.cs
@@ -59,7 +59,6 @@ namespace Server
                         {
                             game.AllPlayerUsePassiveSkill();
                             ReportGame(GameState.GameStart);
-                            game.AllActiveSkillDisabledTemporarily();
                             flag = false;
                         }
                         else ReportGame(GameState.GameRunning);
