+++ logic/规则Logic.md
@@ -196,8 +196,6 @@
 - 不加说明，这里“学生”往往包括职业“教师”
 
 ### 初始状态
-- 所有玩家可以立刻使用主动技能
-  - 虽然一开始会接到仍需的冷却时间，但实际上是CD
 - 玩家出生点固定且一定为空地
 
 ### 交互
